Fixes and changes to Ball Class. Changed most ints to floats to give nicer numbers.

+++ src/ball.h
@@ -25,7 +25,7 @@ public:
     void draw(ofParameterGroup& palette);
     
     // Properties
-    int pos;
+    float pos;
     float radius;
     float pulse;
     bool pulseUp ;
